+++ src/generic.rs
@@ -21,6 +21,8 @@
 //! It is permissible to implement such conversions for multiple [`Rounding`]
 //! modes, for example an [`Approx`] conversion (which rounds inputs where
 //! required) and an [`Exact`] conversion (which rejects these inputs).
+//!
+//! [`RoundInto`]: crate::RoundInto
 
 use crate::{Error, RangeError};
 use core::convert::Infallible;
@@ -146,45 +148,5 @@ impl<S, T: ConvertExact<S>> Convert<S, Approx> for T {
     #[inline]
     fn convert(s: S) -> Self {
         T::convert(s)
-    }
-}
-
-/// Generic "into" conversion trait
-///
-/// This trait is like [`Into`] but for [`Convert`].
-///
-/// The [`Rounding`] mode must be specified when calling this trait's methods,
-/// for example `x.try_round(Exact)` or `y.round(Approx)`. In generic code
-/// (where `R: Rounding`), `z.try_round(R::default())` may be used.
-pub trait RoundInto<T, R: Rounding>: Sized {
-    /// Conversion error type
-    type Error: Into<R::MaximumError> + core::error::Error;
-
-    /// Try converting from `Self` to `T`
-    fn try_round(self, mode: R) -> Result<T, Self::Error>;
-
-    /// Convert from `Self` to `T`
-    ///
-    /// This method must return the same result as [`Self::try_round`] where
-    /// that method succeeds, but differs in the handling of errors:
-    ///
-    /// -   In debug builds the method must panic on error
-    /// -   In release builds the method may return a different value so long as
-    ///     the behaviour is well defined. This allows implementations to
-    ///     optimize to [`as` numeric casts].
-    ///
-    /// [`as` numeric casts]: https://doc.rust-lang.org/reference/expressions/operator-expr.html#type-cast-expressions
-    fn round(self, mode: R) -> T;
-}
-
-impl<R: Rounding, S, T: Convert<S, R>> RoundInto<T, R> for S {
-    type Error = T::Error;
-
-    fn try_round(self, _: R) -> Result<T, Self::Error> {
-        T::try_convert(self)
-    }
-
-    fn round(self, _: R) -> T {
-        T::convert(self)
     }
 }
